requirements.txt
Flask==3.0.0
requests==2.28.2
annotated-types==0.6.0
anyio==4.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
exceptiongroup==1.2.0
Flask-Dance==7.0.0
Flask-SQLAlchemy==3.1.1
gotrue==2.1.0
greenlet==3.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.2
postgrest==0.13.0
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==1.0.2
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
sql-alchemy-db==0.0.4
SQLAlchemy==2.0.23
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.0
supafunc==0.3.3
typing_extensions==4.9.0
urllib3==1.26.18
URLObject==2.4.3
websockets==11.0.3
Werkzeug==3.0.1
zipp==3.17.0